# gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ForgotPasswordErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Send Password Reset Email!
        type: string
    type: object
  models.ForgotPasswordPayload:
    properties:
      identifier:
        example: johndoe/johndoe@example.com
        type: string
    required:
    - identifier
    type: object
  models.ForgotPasswordSuccessResponse:
    properties:
      message:
        example: Password Reset Email Sent Successfully!
        type: string
    type: object
...
  models.ResendActivationEmailErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Send Activation Email!
        type: string
    type: object
  models.ResendActivationEmailPayload:
    properties:
      identifier:
        example: johndoe/johndoe@example.com
        type: string
    required:
    - identifier
    type: object
  models.ResendActivationEmailSuccessResponse:
    properties:
      message:
        example: Activation Email Sent Successfully!
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.ResetPasswordErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Reset Password!
        type: string
    type: object
  models.ResetPasswordPayload:
    properties:
      confirm_password:
        example: strong@password
        type: string
      password:
        example: strong@password
        minLength: 8
        type: string
    required:
    - confirm_password
    - password
    type: object
  models.ResetPasswordSuccessResponse:
    properties:
      message:
        example: Password Reset Successfully!
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        example: "2021-07-01T00:00:00Z"
        type: string
      email:
        example: johndoe@example.com
        type: string
      id:
        example: 67adca0b4e864cfffb002299
        type: string
      is_active:
        example: false
        type: boolean
      role:
        example: customer
        type: string
      updated_at:
        example: "2021-07-01T00:00:00Z"
        type: string
      username:
        example: johndoe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: rohit.vilas.ingole@gmail.com
    name: Rohit Vilas Ingole
  description: This is the API for Gopher Social, a social media platform for Gophers.
  license:
    name: MIT License
    url: https://github.com/DataRohit/Order-Management-System/blob/master/license
  title: Gopher Social API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Forgot Password for a User
      parameters:
      - description: Forgot Password Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully sent password reset email
          schema:
            $ref: '#/definitions/models.ForgotPasswordSuccessResponse'
        "400":
          description: Failed to send password reset email
          schema:
            $ref: '#/definitions/models.ForgotPasswordErrorResponse'
        "500":
          description: Failed to send password reset email
          schema:
            $ref: '#/definitions/models.ForgotPasswordErrorResponse'
      summary: Forgot Password
      tags:
      - Auth
...
  /auth/resend-activation-email:
    post:
      consumes:
      - application/json
      description: Resend Activation Email to a User
      parameters:
      - description: Resend Activation Email Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ResendActivationEmailPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully resent activation email
          schema:
            $ref: '#/definitions/models.ResendActivationEmailSuccessResponse'
        "400":
          description: Failed to resend activation email
          schema:
            $ref: '#/definitions/models.ResendActivationEmailErrorResponse'
        "500":
          description: Failed to resend activation email
          schema:
            $ref: '#/definitions/models.ResendActivationEmailErrorResponse'
      summary: Resend Activation Email
      tags:
      - Auth
  /auth/reset-password/{passwordResetToken}:
    post:
      consumes:
      - application/json
      description: Reset Password for a User
      parameters:
      - description: Password Reset Token
        in: path
        name: passwordResetToken
        required: true
        type: string
      - description: Reset Password Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully reset password
          schema:
            $ref: '#/definitions/models.ResetPasswordSuccessResponse'
        "400":
          description: Failed to reset password
          schema:
            $ref: '#/definitions/models.ResetPasswordErrorResponse'
        "500":
          description: Failed to reset password
          schema:
            $ref: '#/definitions/models.ResetPasswordErrorResponse'
      summary: Reset Password
      tags:
      - Auth
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"